DeleteQuery development
