+++ angular-cc-1-server/server.js
@@ -4,6 +4,8 @@ const cors = require("cors");
 const selectQuery = require('./queries/SelectQuery');
 const InsertQuery = require("./queries/InsertQuery");
 const UpdateQuery = require("./queries/UpdateQuery");
+const DeleteQuery = require("./queries/DeleteQuery");
+
 const app = express();
 const port = 3000;
 
@@ -102,36 +104,22 @@ app.put("/clothes/:id", (req, res) => {
 app.delete("/clothes/:id", (req, res) => {
   const id = parseInt(req.params.id);
 
-  fs.readFile("db.json", "utf8", (err, data) => {
-    if (err) {
-      console.log(err);
-      res.status(500).send("Internal Server Error");
-      return;
-    }
+  DeleteQuery(id)
+      .then(()=>{
+        res.status(204).json({ message: 'Product deleted successfully' });
+      })
+      .catch(
+          err=>{
+            console.error('Error deleting product:', err);
+            res.status(500).json(
+                {
+                  message: 'Failed to delete product'
+                }
+            );
+          }
+      )
 
-    const jsonData = JSON.parse(data);
-
-    const index = jsonData.items.findIndex((item) => item.id === id);
-
-    if (index === -1) {
-      res.status(404).send("Not Found");
-      return;
-    }
-
-    jsonData.items.splice(index, 1);
-
-    fs.writeFile("db.json", JSON.stringify(jsonData), (err) => {
-      if (err) {
-        console.log(err);
-        res.status(500).send("Internal Server Error");
-        return;
-      }
-
-      res.status(204).send();
-    });
-  });
 });
-
 app.listen(port, () => {
   console.log(`Server listening at http://localhost:${port}`);
 });
